refactor: move headers to standard include/ and src/ structure
- Separate public API headers (include/vulkan_game/) from private module headers (src/)
- Update include paths and CMakeLists.txt for new organization
- Improve encapsulation and follow C project conventions
