+++ src/input/mouse.c
@@ -1,5 +1,5 @@
 #include "glfw_fun.h"
-#include "utils.h"
+#include "vulkan_game/utils.h"
 
 static float pitch = 0.0f;
 static double lastX, lastY;
