+++ src/vk/vk_surface.c
@@ -1,5 +1,5 @@
 #include "vk_fun.h"
-#include "glfw_fun.h"
+#include "../input/glfw_fun.h"
 
 GLFWwindow *createVulkanWindow(const char *title){
 	glfwWindowHint(GLFW_CLIENT_API,GLFW_NO_API);
